+++ packages/block-editor/src/utils/block-bindings.js
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { useDispatch, useRegistry } from '@wordpress/data';
-import { getBlockType } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -12,50 +11,6 @@ import { useBlockEditContext } from '../components/block-edit';
 
 function isObjectEmpty( object ) {
 	return ! object || Object.keys( object ).length === 0;
-}
-
-export const BLOCK_BINDINGS_ALLOWED_BLOCKS = {
-	'core/paragraph': [ 'content' ],
-	'core/heading': [ 'content' ],
-	'core/image': [ 'id', 'url', 'title', 'alt' ],
-	'core/button': [ 'url', 'text', 'linkTarget', 'rel' ],
-};
-
-/**
- * Based on the given block name,
- * check if it is possible to bind the block.
- *
- * @param {string} blockName - The block name.
- * @return {boolean} Whether it is possible to bind the block to sources.
- */
-export function canBindBlock( blockName ) {
-	return blockName in BLOCK_BINDINGS_ALLOWED_BLOCKS;
-}
-
-/**
- * Based on the given block name and attribute name,
- * check if it is possible to bind the block attribute.
- *
- * @param {string} blockName     - The block name.
- * @param {string} attributeName - The attribute name.
- * @return {boolean} Whether it is possible to bind the block attribute.
- */
-export function canBindAttribute( blockName, attributeName ) {
-	return (
-		canBindBlock( blockName ) &&
-		BLOCK_BINDINGS_ALLOWED_BLOCKS[ blockName ].includes( attributeName )
-	);
-}
-
-export function getBindableAttributes( blockName ) {
-	const blockType = getBlockType( blockName );
-	const allowedAttributes = BLOCK_BINDINGS_ALLOWED_BLOCKS[ blockName ] || [];
-	// Still comparing with the allowed blocks constant until we can automate the bindings server process.
-	return Object.keys( blockType.attributes ).filter(
-		( key ) =>
-			blockType.attributes[ key ].role === 'content' &&
-			allowedAttributes.includes( key )
-	);
 }
 
 /**
